requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Django==4.1.7
djangorestframework==3.14.0
executing==1.2.0
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
ipython==8.11.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.14.0
python-dotenv==0.7.0
pytz==2023.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.2
text-unidecode==1.3
traitlets==5.9.0
tzdata==2023.2
wcwidth==0.2.6
